New components and path changes
Added files which will be needed for the Shift View.
Modified PageRoutes.jsx to include a pathway to the shift view.
Modified Day.jsx to add useNavigate Hook to clickable days.
These days now redirect to '/master-schedule/shift', but do not pass any data to the new page. This will need to be implemented before final merge.