Add elements to ShiftViewToolbarTop
Added MUI elements to ShiftView's top toolbar.
Add Shift button will redirect to the Time Block editor. Since that page does not exist yet, the button is disabled.
Filter button will contain a list of filter options for all shifts on that day. Filter button's primary color matches the navigation bar's background color. This button will be enabled once that functionality is implemented. It may be removed entirely if it cannot be completed before the end of CS-458.
Close button currently navigates back to /master-schedule. There are plans to add a confirmation modal to the page if changes are made.