Added employee assignment menu
Lots of changes in this one.
Added [unassigned] option to employee list. This will always be the first option in the employee list, even once the process is automated, as it will regulate the enabling and disabling of the Unassign button.
Tooltip on Unassign button was causing errors, so it was removed temporarily. Need to wrap the element in a span.
Added functionality to Unassign button to control button enabling/disabling. Any time the [unassigned] element is no longer selected, the button will enable and vice versa. onClick has been disabled as it crashes the page. Check notes for additional information.
Added employee assignment menu. This menu maps employees to menu items and can be populated with any number of employees available for that shift. There are a few changes that still need to be made to it, but the core functionality is there.
Cleaned up a little formatting and deleted some old code.