Lots of changes
Disabled text wrapping for the text at the left and for the employee assignment menu button. Added a few new names to the dummy data list to demonstrate how the assignment menu handles long names.
Fixed the error the onClick function was causing by wrapping the function call in an arrow function.
Keeping the [unassigned] option enabled in the assignment menu for ease of use. If a way can be found to remove this option from the menu while still having it accessible via the Unassign button, that might be preferable.
Removed TODO tasks which were completed.