Master View refactor complete
Date is now passed all the way down from the parent page to the child Day components.
WeekChanger component passes data back up to MasterView page via a callback function.
I even fixed a bug where the week range would be a week ahead of the desired range whenever the day of the week was Sunday.
Prop validation and defaulting is also complete.