Prop validation beginnings, confirmation dialog
Added a confirmation dialog to the Delete button, asking the user if they want to delete the time block. Dialog also confirms they are deleting the time block, not the position itself.
Changed Delete icon to AutoDelete icon to make it clear that the user is deleting a time block.
Started the beginnings of prop validation.